Update README.md
Task 1: Deploy a simple v1 guestbook application. 

Task 2: Use an in-memory data store for the simple guestbook application. 

Task 3: Update the simple guestbook homepage to include your name. 

Task 4: Automatically deploy the homepage update using a second image stream tag. 

Task 5: Deploy the second version of the guestbook application using an OpenShift build. 

Task 6: Deploy a Redis master, a Redis slave, and an analyzer microservice.

Task 7: Use Redis for the v2 guestbook application instead of an in-memory datastore. 

Task 8: Submit entries to the guestbook and have their tone analyzed. Some simple sentences will not have a tone detected. Ensure that you submit something complex enough so that its tone is detected.

Task 9: Create a Horizontal Pod Autoscaler that shows guestbook as the scale target, the current and desired replicas as three, and the last scale time as the time the deployment was scaled up to three replicas.